requirements.txt
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dash==3.0.4
dash-bootstrap-components==2.0.2
Flask==3.0.3
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
MarkupSafe==3.0.2
narwhals==1.39.1
nest-asyncio==1.6.0
numpy==2.2.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotly==6.1.0
protobuf==6.31.0
pyarrow==20.0.0
pydeck==0.9.1
python-bcb==0.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.25.0
scipy==1.15.3
setuptools==80.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.4
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.0.6
zipp==3.21.0
